pyproject.toml
[tool.poetry]
name = "ioclcd"
version = "0.1.0"
description = "iocurrents LCD display"
authors = ["Rob Bogdan <rob@iocurrents.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyusb = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
